Made data/utils module to not have viv_inputs in build dependencies.  Updated SSN generator due to inheritance bug
